Update entity package to common

+++ src/main/java/com/cardano/explorer/controller/BlockController.java
@@ -29,7 +29,7 @@ public class BlockController {
   @LogMessage
   @Operation(summary = "Get a block detail by its no")
   public BlockResponse getBlockDetail(
-      @PathVariable @Parameter(description = "Block number") Integer no) {
+      @PathVariable @Parameter(description = "Block number") Long no) {
     return blockService.getBlockDetail(no);
   }
